agent_service/requirements.txt
fastapi
uvicorn
PyGithub
python-dotenv
requests
google-genai
